Removed IDatabase and moved facade
