--- src/main/java/com/web/model/database/IDatabase.java
@@ -1,18 +1,0 @@
-package com.web.model.database;
-
-import java.util.Map;
-
-import com.web.model.Book;
-
-/**
- * Database interface to handle database operations for the book library
- * 
- */
-public interface IDatabase {
-
-	public Map<Integer, String> listBooks();
-
-	public Book getBook(int id);
-
-	public void addBook(Book book);
-}
